requirements.txt
# Core API framework
fastapi==0.115.0
uvicorn==0.30.6

# Deep Learning
torch
torchvision
facenet-pytorch==2.5.3
efficientnet-pytorch==0.7.1  # (optional, safe to keep)

# Image / Video processing
Pillow==10.2.0
opencv-python==4.10.0.84

# Utilities
tqdm==4.66.5
numpy==1.26.4

# Optional but useful
python-multipart==0.0.9   # Enables file uploads in FastAPI
